requirements.txt
faiss-cpu==1.7.4
python-liquid==1.10.2
openai==0.28.0
langchain==0.0.345
tqdm==4.66.1
transformers==4.37.2
sentence_transformers==2.2.2
# pyserini==0.22.1
tiktoken==0.6.0
datasets==2.16.1
annotated-types==0.7.0
anyio==4.7.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0